VMS-55487: Enable -Werror for all general purpose libraries by default

+++ libs/nx_codec/src/nx/codec/h265/short_term_reference_picture_set.cpp
@@ -86,7 +86,7 @@ bool ShortTermReferencePictureSet::read(
         for (uint32_t i = 0; i < DeltaPocS0.size(); ++i)
         {
             if (i == 0)
-                DeltaPocS0[i] = -(delta_poc_s0_minus1_vector[i] + 1);
+                DeltaPocS0[i] = -((int32_t) delta_poc_s0_minus1_vector[i] + 1);
             else
                 DeltaPocS0[i] = delta_poc_s0_minus1_vector[i - 1] - (delta_poc_s0_minus1_vector[i] + 1);
         }
